requirements.txt
a2wsgi==1.10.10
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.1.8
apache-airflow-core==3.1.8
apache-airflow-providers-common-compat==1.14.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.32.0
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.12.0
apache-airflow-task-sdk==1.1.8
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
attrs==25.4.0
babel==2.18.0
blinker==1.9.0
cachetools==7.0.5
cadwyn==5.4.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.98.0
datasets==4.7.0
Deprecated==1.3.1
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
dynaconf==3.2.12
email-validator==2.3.0
evaluate==0.4.6
evidently==0.7.21
Faker==40.8.0
fastapi==0.117.1
fastapi-cli==0.0.24
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
googleapis-common-protos==1.73.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenback==1.3.0
greenlet==3.3.2
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
litestar==2.21.1
litestar-htmx==0.5.0
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
methodtools==0.4.7
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.18.0
natsort==8.4.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
patsy==1.0.2
pendulum==3.2.0
pillow==12.1.1
plotly==5.24.1
pluggy==1.6.0
polyfactory==3.3.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.11.0
pyparsing==3.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
PyYAML-ft==8.0.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-argparse==1.7.2
rich-click==1.9.7
rich-toolkit==0.19.7
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setproctitle==1.3.7
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
statsmodels==0.14.6
streamlit==1.55.0
structlog==25.5.0
svcs==25.1.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
ujson==5.12.0
universal_pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.41.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
wirerope==1.0.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
